Update elf dependency to 0.7.1
This pulls in a bunch of interface changes to the elf crate. Most of the
changes in the diff are just from some type names changing.

The manual parsing of the build-id note is no longer needed, as the elf crate can parse
various note formats now.

The ElfBytes type is a partially zero-alloc and lazy-parsing type, so I read the file data once
and stashed that in an allocated Vec on the NxoFile then parsed out what was needed to try and keep
the diff as simple as possible :)